backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0
numpy==1.26.2
scikit-learn==1.3.2
pandas==2.1.4
gunicorn==21.2.0
